Move stagemaster into separate crate
