+++ src/core/mod.rs
@@ -1,5 +1,4 @@
-pub mod geometry;
+pub mod colors;
 pub mod simulation;
 pub mod disjoint_sets;
 pub mod grid_accelerator;
-pub mod stagemaster;
