+++ src/game/lanes_and_cars/planning/current_plan/selectable.rs
@@ -1,7 +1,7 @@
 use kay::{ID, Recipient, Actor, Fate};
 use kay::swarm::{Swarm, SubActor, CreateWith};
 use descartes::{N, Band, Curve, Into2d, FiniteCurve, Path, RoughlyComparable};
-use core::geometry::{CPath, AnyShape};
+use stagemaster::geometry::{CPath, AnyShape};
 
 use super::{SelectableStrokeRef, CurrentPlan};
 
@@ -23,7 +23,7 @@ impl Selectable {
 }
 
 use super::InitInteractable;
-use core::stagemaster::{UserInterface, AddInteractable};
+use stagemaster::{UserInterface, AddInteractable};
 
 impl Recipient<InitInteractable> for Selectable {
     fn receive(&mut self, _msg: &InitInteractable) -> Fate {
@@ -36,7 +36,7 @@ impl Recipient<InitInteractable> for Selectable {
 }
 
 use super::ClearInteractable;
-use core::stagemaster::RemoveInteractable;
+use stagemaster::RemoveInteractable;
 
 impl Recipient<ClearInteractable> for Selectable {
     fn receive(&mut self, _msg: &ClearInteractable) -> Fate {
@@ -45,7 +45,7 @@ impl Recipient<ClearInteractable> for Selectable {
     }
 }
 
-use core::stagemaster::Event3d;
+use stagemaster::Event3d;
 use super::{ChangeIntent, Intent, IntentProgress, ContinuationMode};
 
 const START_END_SNAP_DISTANCE: N = 10.0;
